requirements.txt
langchain>=0.3.0
langchain-core>=0.3.68
langchain-openai>=0.3.28
langchain-google-genai>=2.1.9
langgraph>=0.0.20
openai>=1.99.1
pydantic>=2.11.7
python-dotenv>=1.0.0
pyodbc>=5.2.0 